add BoundingBox tests and remove bounding_box
