+++ tests/geometry/test_bound_box.py
@@ -1,0 +1,45 @@
+# bluemira is an integrated inter-disciplinary design tool for future fusion
+# reactors. It incorporates several modules, some of which rely on other
+# codes, to carry out a range of typical conceptual fusion reactor design
+# activities.
+#
+# Copyright (C) 2022 M. Coleman, J. Cook, F. Franza, I.A. Maione, S. McIntosh, J. Morris,
+#                    D. Short
+#
+# bluemira is free software; you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public
+# License as published by the Free Software Foundation; either
+# version 2.1 of the License, or (at your option) any later version.
+#
+# bluemira is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public
+# License along with bluemira; if not, see <https://www.gnu.org/licenses/>.
+
+import numpy as np
+
+from bluemira.geometry.bound_box import BoundingBox
+
+
+class TestBoundingBox:
+    def test_null(self):
+        x, y, z = np.zeros(100), np.zeros(100), np.zeros(100)
+        xb, yb, zb = BoundingBox.from_xyz(x, y, z).get_box_arrays()
+        assert np.all(xb == 0)
+        assert np.all(yb == 0)
+        assert np.all(zb == 0)
+
+    def test_random(self):
+        x, y, z = np.random.rand(100), np.random.rand(100), np.random.rand(100)
+        args = np.random.randint(0, 100, 8)
+        x[args] = np.array([-2, -2, -2, -2, 2, 2, 2, 2])
+        y[args] = np.array([-2, -2, 2, 2, 2, -2, -2, 2])
+        z[args] = np.array([-2, 2, -2, 2, -2, 2, -2, 2])
+        xb, yb, zb = BoundingBox.from_xyz(x, y, z).get_box_arrays()
+
+        assert np.allclose(xb, np.array([-2, -2, -2, -2, 2, 2, 2, 2]))
+        assert np.allclose(yb, np.array([-2, -2, 2, 2, -2, -2, 2, 2]))
+        assert np.allclose(zb, np.array([-2, 2, -2, 2, -2, 2, -2, 2]))
